# docs/swagger.yaml
basePath: /api/v1
definitions:
  TodoApp_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      description:
        example: Сделать до понедельника домашнее задание
        maxLength: 1000
        minLength: 1
        type: string
      title:
        example: Домашнее задание
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d468
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до понедельника домашнее задание
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      title:
        example: Домашнее задание
        type: string
      version:
        example: 2
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d468
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до понедельника домашнее задание
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      title:
        example: Домашнее задание
        type: string
      version:
        example: 2
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        example: true
        type: boolean
      description:
        example: Новое описание
        type: string
      title:
        example: Новое название
        type: string
    type: object
  internal_features_tasks_transport_http.PatchTaskResponse:
    properties:
      author_user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d468
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до понедельника домашнее задание
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      title:
        example: Домашнее задание
        type: string
      version:
        example: 2
        type: integer
    type: object
  internal_features_tasks_transport_http.TasksDTOResponse:
    properties:
      author_user_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d468
        type: string
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до понедельника домашнее задание
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      title:
        example: Домашнее задание
        type: string
      version:
        example: 2
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivan Ivanov
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+78005353535"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: f25b2127-1f7a-40c8-bbca-d3c766e99565
        type: string
      phone_number:
        example: "78005353535"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: f25b2127-1f7a-40c8-bbca-d3c766e99565
        type: string
      phone_number:
        example: "78005353535"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Денис Денисоввич
        type: string
      phone_number:
        example: "+78005553535"
        type: string
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: f25b2127-1f7a-40c8-bbca-d3c766e99565
        type: string
      phone_number:
        example: "78005353535"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: f25b2127-1f7a-40c8-bbca-d3c766e99565
        type: string
      phone_number:
        example: "78005353535"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
  /tasks:
    get:
      description: Просмотр списка задач с опциональной пагинацией и/или фильтрацие
        по ID автора задачи
      parameters:
      - description: Размер страницы с задачами
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с задачами
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список задач
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TasksDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу в системе
      parameters:
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданная задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Удаление существующей задачи в системе по ее ID
      responses:
        "204":
          description: Успешное удаление задачи
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление задачи
      tags:
      - tasks
    get:
      description: Получение конкретной задачи по ID
      produces:
      - application/json
      responses:
        "200":
          description: Задача успешно найдена
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение задачи
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующей в системе задачи
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: `description` игнорируется, значение в бд не меняется
        1. **Явно передано значение**: `"description": "Утром в 6:30 выйти на пробежку" - устанавливает новое описание для задачи`
        1. **Передан null**: `"description": null` - очищает поле в БД (set to NULL)
        1. Ограничение: `title` и `completed` не могут быть выставлены как null
      parameters:
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённая задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: изменение задачи
      tags:
      - tasks
  /users:
    get:
      description: Просмотр списка пользователей с опциональной пагинацией
      parameters:
      - description: Размер страницы с пользователями
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с пользователями
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка пользователей
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаление существующего в системе пользователя по его ID
      responses:
        "204":
          description: Успешное удаление пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление пользователя
      tags:
      - users
    get:
      description: Получение конкретного пользователя по ID
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно найден
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующем в системе пользователе
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: `phone_number` игнорируется, значение в бд не меняется
        2. **Явно передано значение**: `"phone_number": "+78005553535"` - устанавливает новый номер телефона в бд
        3. **Передан null**: `"phone_number": null` - очищает поле в БД (set to NULL)
        4. Ограничение: `full_name` не может быть выставлен как null
      parameters:
      - description: PatchUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённый пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: изменение пользователя
      tags:
      - users
swagger: "2.0"